requirement.txt
monai==1.1.0
numpy==1.19.5
Pillow==8.4.0
pip==19.2.3
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
timm==0.9.12
torch==1.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.36.1
typing_extensions==4.0.1

